API/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
mysql-connector-python==9.4.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44